refactor: move software cursor helpers into dedicated sw_cursor.{c,h}
Extract the swCursor struct and its five helpers (Create, Move, Restamp,
Destroy, GetBgWin) out of main.c into src/sw_cursor.c / src/sw_cursor.h.
The struct is opaque in the header; the #if ENABLE_XRENDER guard lives
inside sw_cursor.c alongside the no-op stubs for non-XRender builds.
sw_cursor.c is added to alock_SOURCES unconditionally — no new Makefile
conditional is needed.  main.c now simply includes sw_cursor.h and no
longer pulls in <X11/extensions/Xrender.h> directly.

Addresses review feedback on PR #24 requesting these helpers be kept
outside of the main.c compilation unit.
